fix: make router fast
There is no need to create goroutines in the router when
we can instead let the Link's writer do a bit more work on
behalf of the router. Additionally, reckon with the fact
that we need a queue of notifications.

With this change implemented, `calibrate -star` runs at
around 1.5 Gbit/s, while `calibrate` runs at 1.6. This
drop in performance is acceptable, and definitely a bit
better than the previous speed (5 Mbit/s).