fix(calibrate): produce actionable CSV entries (#5)
We need to include in each row:

1. the prefix of the PCAP files we're capturing;

2. the imposed RTT;

3. the imposed packet loss rate.

We need to also capture packets at the client, because otherwise we
cannot observe whether we're producing any traffic bursts.

We need to refactor the client measurement algorithm such that we
include a final sample with the overall average goodput. To this end, we
include a boolean telling us whether an individual CSV row is an interim
or a final sample. Data consumers could then grep for that flag to only
collect the final sample.

While there:

1. ignore more files in `.gitignore`;

2. make sure `calibrate` is not stuck in accept by explicitly closing
the listener.