feat: introduce a delay-only forwarding algorithm
We will select this algorithm when the user only configures a
delay without any PLR or DPI need.

This link causes the TCP sender to be receiver limited, so
the TCP performance we get with losses are very bad.

We will need to introduce a better link that can emulate also
PLR and DPI (which MAY cause some packet losses).