[API CHANGE] feat: introduce framework for spoofing packets
This diff generalizes the code we previously used to send a RST
segment when we see an offending SNI. Rather than having a specific
flag for reflecting a TCP segment with the RST flag set, we change
the code to allow DPI rules to tell the router it should spoof zero
or more packets. The DPI rule determines the spoofed packets and
the router just tries to route each spoofed packet.

After this change, integration testing for RST segments show that we
are still able to spoof back TCP segments with the RST flag for specific
offending SNIs. However, now the code is more general and allows us to
parse DNS requests and spoof DNS replies.