feat: implement string-based TCP RST
This diff adds support for sending a RST segment when the client to
server flow includes a specific string in the first few packets.

We're mainly going to use this functionality to RST cleartext HTTP
requests containing a specific host header.

I wrote this diff for https://github.com/ooni/probe/issues/2494. I need
a test case in which an HTTP round trip fails, and this diff seems the
easiest way to implement this functionality quickly. What's more, there
are several censorship devices behaving exactly like this.