feat(dpiblock.go): implement string-based TCP RST (#29)
This diff adds support for sending a RST segment when the client to
server flow includes a specific string in the first few packets and the
server endpoint matches.

We're mainly going to use this functionality to RST cleartext HTTP
requests containing a specific host header.

I wrote this diff for https://github.com/ooni/probe/issues/2494. I need
a test case in which an HTTP round trip fails, and this diff seems the
easiest way to implement this functionality quickly. What's more, this
is a common censorship case.

While there, use `crypto/rand` instead of `math/rand` to avoid a
warning.