feat: make sure DPI does not affect loopback traffic (#39)
It should not happen because GVisor should route the traffic to itself
but it's better to be sure about this property with a test.

I was wondering about whether this property hold when I was reasoning
about netemx tests I need for https://github.com/ooni/probe/issues/2531.

While there, refrain from capturing PCAPs in examples. This was a
leftover from when I was debugging the examples.