BREAKING CHANGE: fix: pin certificates to hosts
Before writing this diff, netem automatically generated the correct
cert on the fly based on the given SNI or local addr.

While this behavior has been okay so far, it turns out there is a
set of tests we cannot write because of it.

Namely, we cannot check whether we can connect to a given host
using another SNI, because netem would generate a certificate
for the provided SNI, which is not how the internet works.

If I'm connecting to www.example.com with www.google.com as the
SNI, in the internet the server would return a valid certificate
for www.example.com, rather than for www.google.com.

This diff rectifies netem's behavior by forcing the user to pin
the certificate to a set of names and addresses when creating the
server that needs such a certificate.

Accordingly, we can stop using a forked google/martian/v3/mitm
implementation and we can just roll out our own, which is still
based on martian (feat not, not reinventing the wheel here)
but allows us to create a certificate with specific addresses
and domain names pinned to it.

While there, notice that there was code we were not using,
such as stdlib.go, and that we also don't need in ooni/probe-cli,
so we can definitely kill this piece of code.

Part of https://github.com/ooni/probe/issues/2531, because the
tests I was trying to write belong to such an issue.