fix(UNetStack): better compatibility with ooni/probe-cli models (#41)
The refactoring at
https://github.com/ooni/netem/commit/45a7aa975636a8de271e07904ccd076a953618b0
was a bit too aggressive and we have lost some degree of interface level
compatibility with ooni/probe-cli. Let's repair this:

* Add the `CACert` accessor required by ooni/probe-cli code.

* While there, also declare the accessor as being part of
`UnderlyingNetwork`.

* We have a slight issue with `ServerTLSConfig`, which is expected by
ooni/probe-cli and now is called `MustNewTLSConfig` and, additionally,
requires the names for which to generate the certificate. We will solve
these issues inside the ooni/probe-cli repository, but let's make sure
the `UNetStack` has the correct method.

Part of https://github.com/ooni/probe/issues/2531