BREAKING CHANGE: unify UNetStack and CA's CertificationAuthority behavior
The UNetStack and CA structures have similar functionality exposed
through a different interface, which causes us to create an
interface and a wrapper for the code in probe-cli.

I noticed about this when I was trying to integrate new netem
changes, and it occurred to me we can just make the interfaces
uniform here and avoid wrapping in probe-cli.

Also part of https://github.com/ooni/probe/issues/2531